+++ packages/axway-cli-oum/CHANGELOG.md
@@ -1,4 +1,8 @@
-# v2.0.1 (Oct 18, 2021)
+# v2.0.2 (Nov 5, 2021)
+
+ * chore: Updated dependencies.
+
+* # v2.0.1 (Oct 18, 2021)
 
  * chore: Updated dependencies.